+++ src/components/antd/main.js
@@ -6,3 +6,4 @@ export * from './Table';
 export * from './Timeline';
 export * from './MenuItem';
 export * from './Collapse';
+export * from './SimpleProgressBar';