+++ src/components/manager/main.js
@@ -3,3 +3,4 @@ export { createChart } from './ChartManager/utils';
 export * from './Responsive';
 export * from './TableManager';
 export * from './ElasticWidget';
+export * from './BreadcrumbManager';