+++ src/helpers/number.js
@@ -1,0 +1,4 @@
+export const roundToN = d => {
+  const multiplier = Math.pow(10, d);
+  return num => Math.floor(num * multiplier) / multiplier;
+};